requirements.txt
openai
faiss-cpu
sentence-transformers
numpy
python-dotenv # for managing OpenAI keys
